- Continued integrating SEEDS API with the Avatar/Karma System.
- Added SEEDS API to OASIS API REST Service:

Added the following methods:

- GetAllOrganisations
- PayWithSeeds
- PayAvatarWithSeeds
- RewardWithSeeds
- RewardAvatarWithSeeds
- DonateWithSeeds
- DonateAvatarWithSeeds
- SendInviteToJoinSeeds
- SendInviteToAvatarJoinSeeds
- AcceptInviteToJoinSeeds
- GetTelosAccountNameForAvatar
- GetTelosAccount
- GetTelosAccountForAvatar
- GetAvatarIdForTelosAccountName
- GetAvatarForTelosAccountName
- GetBalance
- GetBalanceForAvatar
- GenerateSeedsPassportSignInQRCode
- GenerateSeedsPassportSignInQRCodeForAvatar
- AddKarmaForSeeds

+++ NextGenSoftware.OASIS.API.Providers.SEEDSOASIS/ParamObjects/Invite.cs
@@ -6,7 +6,7 @@ namespace NextGenSoftware.OASIS.API.Providers.SEEDSOASIS.ParamObjects
         public string sponsor { get; set; }
         public string referrer { get; set; }
         public string invite_hash { get; set; }
-        public int transfer_quantity { get; set; }
-        public int sow_quantity { get; set; }
+        public float transfer_quantity { get; set; }
+        public float sow_quantity { get; set; }
     }
 }
